pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "pypdf2>=3.0.1",
    "requests>=2.32.3",
    "sift-stack-py>=0.4.2",
]
